Prospective fix for preset layout transition issues
