+++ Vocable/Features/Settings/EditSayings/EditSayingsKeyboardViewController.swift
@@ -11,7 +11,7 @@ import AVFoundation
 import UIKit
 import CoreData
 
-class EditSayingsKeyboardViewController: UIViewController, UICollectionViewDelegate {
+class EditSayingsKeyboardViewController: UIViewController, UICollectionViewDelegate, VocableCollectionViewLayoutTransitioningDelegate {
     
     private var dataSource: UICollectionViewDiffableDataSource<Section, ItemWrapper>!
     
@@ -106,16 +106,16 @@ class EditSayingsKeyboardViewController: UIViewController, UICollectionViewDeleg
     }
     
     private func createLayout() -> UICollectionViewLayout {
-        let layout = PresetUICollectionViewCompositionalLayout { (sectionIndex: Int, layoutEnvironment: NSCollectionLayoutEnvironment) -> NSCollectionLayoutSection? in
+        let layout = PresetCollectionViewCompositionalLayout { (sectionIndex: Int, layoutEnvironment: NSCollectionLayoutEnvironment) -> NSCollectionLayoutSection? in
             let sectionKind = self.dataSource.snapshot().sectionIdentifiers[sectionIndex]
             
             switch sectionKind {
             case .textField:
                 return self.topBarLayout()
             case .keyboard:
-                return PresetUICollectionViewCompositionalLayout.keyboardLayout(with: layoutEnvironment)
+                return PresetCollectionViewCompositionalLayout.keyboardLayout(with: layoutEnvironment)
             case .suggestions:
-                return PresetUICollectionViewCompositionalLayout.suggestiveTextSectionLayout(with: layoutEnvironment)
+                return PresetCollectionViewCompositionalLayout.suggestiveTextSectionLayout(with: layoutEnvironment)
             }
         }
         layout.register(CategorySectionBackground.self, forDecorationViewOfKind: "CategorySectionBackground")
